fix: moved sirebar settings to settings store
