+++ frontend/src/stores/session.js
@@ -54,12 +54,6 @@ export const sessionStore = defineStore('lms-session', () => {
 		},
 	})
 
-	const sidebarSettings = createResource({
-		url: 'lms.lms.api.get_sidebar_settings',
-		cache: 'Sidebar Settings',
-		auto: false,
-	})
-
 	const livecodeURL = createResource({
 		url: 'frappe.client.get_single_value',
 		params: {
@@ -77,7 +71,6 @@ export const sessionStore = defineStore('lms-session', () => {
 		logout,
 		brand,
 		branding,
-		sidebarSettings,
 		livecodeURL,
 	}
 })
